sheep/store: convert to dynamic string allocation

+++ sheep/store/tree_store.c
@@ -19,39 +19,53 @@ static inline int get_tree_directory(uint64_t oid)
 	return (int) ((oid & TREE_DIR_MASK) >> VDI_SPACE_SHIFT) ;
 }
 
-static int get_store_path(uint64_t oid, uint8_t ec_index, char *path)
+static int get_store_path(uint64_t oid, uint8_t ec_index, char **path)
 {
-	char tree_path[PATH_MAX - 33];
+	char *tree_path;
+	int ret;
 
 	if (is_vdi_obj(oid) || is_vmstate_obj(oid) || is_vdi_attr_obj(oid)) {
-		snprintf(tree_path, sizeof(tree_path), "%s/meta",
+		ret = asprintf(&tree_path, "%s/meta",
 			 md_get_object_dir(oid));
 	} else {
-		snprintf(tree_path, sizeof(tree_path), "%s/%02x",
+		ret = asprintf(&tree_path, "%s/%02x",
 			 md_get_object_dir(oid), get_tree_directory(oid));
 	}
+	if (ret < 0)
+		return ret;
 
 	if (is_erasure_oid(oid)) {
-		if (unlikely(ec_index >= SD_MAX_COPIES))
+		if (unlikely(ec_index >= SD_MAX_COPIES)) {
 			panic("invalid ec_index %d", ec_index);
-		return snprintf(path, PATH_MAX, "%s/%016"PRIx64"_%d",
+			free(tree_path);
+			errno = EINVAL;
+			return -1;
+		}
+		ret = asprintf(path, "%s/%016"PRIx64"_%d",
 				tree_path, oid, ec_index);
-	}
-
-	return snprintf(path, PATH_MAX, "%s/%016" PRIx64, tree_path, oid);
+	} else
+		ret = asprintf(path, "%s/%016" PRIx64, tree_path, oid);
+	free(tree_path);
+	if (ret < 0)
+		free(path);
+	return ret;
 }
 
-static int get_store_tmp_path(uint64_t oid, uint8_t ec_index, char *path)
+static int get_store_tmp_path(uint64_t oid, uint8_t ec_index, char **path)
 {
-	char tmp_path[PATH_MAX - 4];
+	char *tmp_path;
+	int ret;
 
-	get_store_path(oid, ec_index, path);
-	memcpy(tmp_path, path, sizeof(tmp_path));
-	return snprintf(path, PATH_MAX, "%s.tmp", tmp_path);
+	ret = get_store_path(oid, ec_index, &tmp_path);
+	if (ret < 0)
+		return ret;
+	ret = asprintf(path, "%s.tmp", tmp_path);
+	free(tmp_path);
+	return ret;
 }
 
 static int get_store_stale_path(uint64_t oid, uint32_t epoch, uint8_t ec_index,
-				char *path)
+				char **path)
 {
 	return md_get_stale_path(oid, epoch, ec_index, path);
 }
@@ -62,11 +76,19 @@ static int get_store_stale_path(uint64_t oid, uint32_t epoch, uint8_t ec_index,
  */
 bool tree_exist(uint64_t oid, uint8_t ec_index)
 {
-	char path[PATH_MAX];
+	char *path;
+	int ret;
 
-	get_store_path(oid, ec_index, path);
+	ret = get_store_path(oid, ec_index, &path);
+	if (ret < 0)
+		return false;
 
-	return md_exist(oid, ec_index, path);
+	if (md_exist(oid, ec_index, path)) {
+		free(path);
+		return true;
+	}
+	free(path);
+	return false;
 }
 
 /* Trim zero blocks of the beginning and end of the object. */
@@ -103,7 +125,7 @@ int tree_write(uint64_t oid, const struct siocb *iocb)
 {
 	int flags = prepare_iocb(oid, iocb, false), fd,
 	    ret = SD_RES_SUCCESS;
-	char path[PATH_MAX];
+	char *path;
 	ssize_t size;
 	uint32_t len = iocb->length;
 	uint64_t offset = iocb->offset;
@@ -124,19 +146,27 @@ int tree_write(uint64_t oid, const struct siocb *iocb)
 		sync();
 	}
 
-	get_store_path(oid, iocb->ec_index, path);
+	ret = get_store_path(oid, iocb->ec_index, &path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	/*
 	 * Make sure oid is in the right place because oid might be misplaced
 	 * in a wrong place, due to 'shutdown/restart with less/more disks' or
 	 * any bugs. We need call err_to_sderr() to return EIO if disk is broken
 	 */
-	if (!tree_exist(oid, iocb->ec_index))
-		return err_to_sderr(path, oid, ENOENT);
+	if (!tree_exist(oid, iocb->ec_index)) {
+		ret = err_to_sderr(path, oid, ENOENT);
+		free(path);
+		return ret;
+	}
 
 	fd = open(path, flags, sd_def_fmode);
-	if (unlikely(fd < 0))
-		return err_to_sderr(path, oid, errno);
+	if (unlikely(fd < 0)) {
+		ret = err_to_sderr(path, oid, errno);
+		free(path);
+		return ret;
+	}
 
 	if (trim_is_supported && is_sparse_object(oid)) {
 		if (tree_trim(fd, oid, iocb, &offset, &len) < 0) {
@@ -156,17 +186,22 @@ int tree_write(uint64_t oid, const struct siocb *iocb)
 	}
 out:
 	close(fd);
+	free(path);
 	return ret;
 }
 
 static int make_tree_dir(const char *path)
 {
-	int i;
-	char p[PATH_MAX + 5];
+	int i, ret;
+	char *p;
 
-	snprintf(p, PATH_MAX, "%s/meta", path);
+	ret = asprintf(&p, "%s/meta", path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
+
 	if (xmkdir(p, sd_def_dmode) < 0) {
 		sd_err("%s failed, %m", p);
+		free(p);
 		return SD_RES_EIO;
 	}
 
@@ -174,23 +209,28 @@ static int make_tree_dir(const char *path)
 		snprintf(p, PATH_MAX, "%s/%02x", path, i);
 		if (xmkdir(p, sd_def_dmode) < 0) {
 			sd_err("%s failed, %m", p);
+			free(p);
 			return SD_RES_EIO;
 		}
 	}
-
+	free(p);
 	return SD_RES_SUCCESS;
 }
 
 static int make_stale_dir(const char *path)
 {
-	char p[PATH_MAX];
+	char *p;
+	int ret;
 
-	snprintf(p, PATH_MAX, "%s/.stale", path);
+	ret = asprintf(&p, "%s/.stale", path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 	if (xmkdir(p, sd_def_dmode) < 0) {
 		sd_err("%s failed, %m", p);
+		free(p);
 		return SD_RES_EIO;
 	}
-
+	free(p);
 	return SD_RES_SUCCESS;
 }
 
@@ -204,13 +244,19 @@ static int purge_dir(const char *path)
 
 static int purge_stale_dir(const char *path)
 {
-	char p[PATH_MAX];
+	char *p;
+	int ret;
 
-	snprintf(p, PATH_MAX, "%s/.stale", path);
+	ret = asprintf(&p, "%s/.stale", path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
-	if (purge_directory_async(p) < 0)
+	if (purge_directory_async(p) < 0) {
+		free(p);
 		return SD_RES_EIO;
+	}
 
+	free(p);
 	return SD_RES_SUCCESS;
 }
 
@@ -262,18 +308,23 @@ static int init_objlist_and_vdi_bitmap(uint64_t oid, const char *wd,
 				       void *arg)
 {
 	int ret;
-	char path[PATH_MAX];
+	char *path;
 	objlist_cache_insert(oid);
 
-	snprintf(path, PATH_MAX, "%s/meta", wd);
+	ret = asprintf(&path, "%s/meta", wd);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	if (is_vdi_obj(oid)) {
 		sd_debug("found the VDI object %016" PRIx64" epoch %"PRIu32
 			 " at %s", oid, epoch, path);
 		ret = init_vdi_state(oid, path, epoch);
-		if (ret != SD_RES_SUCCESS)
+		if (ret != SD_RES_SUCCESS) {
+			free(path);
 			return ret;
+		}
 	}
+	free(path);
 	return SD_RES_SUCCESS;
 }
 
@@ -331,9 +382,11 @@ static int tree_read_from_path(uint64_t oid, const char *path,
 int tree_read(uint64_t oid, const struct siocb *iocb)
 {
 	int ret;
-	char path[PATH_MAX];
+	char *path;
 
-	get_store_path(oid, iocb->ec_index, path);
+	ret = get_store_path(oid, iocb->ec_index, &path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 	ret = tree_read_from_path(oid, path, iocb);
 
 	/*
@@ -342,16 +395,20 @@ int tree_read(uint64_t oid, const struct siocb *iocb)
 	 */
 	if (ret == SD_RES_NO_OBJ && iocb->epoch > 0 &&
 	    iocb->epoch < sys_epoch()) {
-		get_store_stale_path(oid, iocb->epoch, iocb->ec_index, path);
+		free(path);
+		ret = get_store_stale_path(oid, iocb->epoch,
+					   iocb->ec_index, &path);
+		if (ret < 0)
+			return SD_RES_NO_MEM;
 		ret = tree_read_from_path(oid, path, iocb);
 	}
-
+	free(path);
 	return ret;
 }
 
 int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 {
-	char path[PATH_MAX], tmp_path[PATH_MAX], *dir;
+	char *path, *tmp_path, *dir;
 	int flags = prepare_iocb(oid, iocb, true);
 	int ret, fd;
 	uint32_t len = iocb->length;
@@ -360,8 +417,14 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 	uint64_t offset = iocb->offset;
 
 	sd_debug("%016"PRIx64, oid);
-	get_store_path(oid, iocb->ec_index, path);
-	get_store_tmp_path(oid, iocb->ec_index, tmp_path);
+	ret = get_store_path(oid, iocb->ec_index, &path);
+	if (ret < 0)
+		return ret;
+	ret = get_store_tmp_path(oid, iocb->ec_index, &tmp_path);
+	if (ret < 0) {
+		free(tmp_path);
+		return ret;
+	}
 
 	if (uatomic_is_true(&sys->use_journal) &&
 	    journal_write_store(oid, iocb->buf, iocb->length,
@@ -384,11 +447,16 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 			 * so it is okay to simply return success here.
 			 */
 			sd_debug("%s exists", tmp_path);
+			free(tmp_path);
+			free(path);
 			return SD_RES_SUCCESS;
 		}
 
 		sd_err("failed to open %s: %m", tmp_path);
-		return err_to_sderr(path, oid, errno);
+		free(tmp_path);
+		ret = err_to_sderr(path, oid, errno);
+		free(path);
+		return ret;
 	}
 
 	obj_size = get_store_objsize(oid);
@@ -404,7 +472,7 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 			ret = prealloc(fd, obj_size);
 		if (ret < 0) {
 			ret = err_to_sderr(path, oid, errno);
-			goto out;
+			goto out_close;
 		}
 	}
 
@@ -412,21 +480,22 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 	if (ret != len) {
 		sd_err("failed to write object. %m");
 		ret = err_to_sderr(path, oid, errno);
-		goto out;
+		goto out_close;
 	}
 
 	ret = rename(tmp_path, path);
 	if (ret < 0) {
 		sd_err("failed to rename %s to %s: %m", tmp_path, path);
 		ret = err_to_sderr(path, oid, errno);
-		goto out;
+		goto out_close;
 	}
 
 	close(fd);
 
 	if (uatomic_is_true(&sys->use_journal) || sys->nosync == true) {
 		objlist_cache_insert(oid);
-		return SD_RES_SUCCESS;
+		ret = SD_RES_SUCCESS;
+		goto out_free;
 	}
 
 	pstrcpy(tmp_path, sizeof(tmp_path), path);
@@ -434,7 +503,9 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 	fd = open(dir, O_DIRECTORY | O_RDONLY);
 	if (fd < 0) {
 		sd_err("failed to open directory %s: %m", dir);
-		return err_to_sderr(path, oid, errno);
+		ret = err_to_sderr(path, oid, errno);
+		goto out_free;
+
 	}
 
 	if (fsync(fd) != 0) {
@@ -443,36 +514,48 @@ int tree_create_and_write(uint64_t oid, const struct siocb *iocb)
 		close(fd);
 		if (unlink(path) != 0)
 			sd_err("failed to unlink %s: %m", path);
-		return ret;
+		goto out_free;
 	}
 	close(fd);
 	objlist_cache_insert(oid);
+	free(tmp_path);
+	free(path);
 	return SD_RES_SUCCESS;
 
-out:
+out_close:
 	if (unlink(tmp_path) != 0)
 		sd_err("failed to unlink %s: %m", tmp_path);
 	close(fd);
+out_free:
+	free(tmp_path);
+	free(path);
 	return ret;
 }
 
 int tree_link(uint64_t oid, uint32_t tgt_epoch)
 {
-	char path[PATH_MAX + 34], stale_path[PATH_MAX], tree_path[PATH_MAX];
+	char *path, *stale_path, *tree_path;
+	int ret;
 
 	if (is_vdi_obj(oid) || is_vmstate_obj(oid) || is_vdi_attr_obj(oid)) {
-		snprintf(tree_path, PATH_MAX, "%s/meta",
+		ret = asprintf(&tree_path, "%s/meta",
 			 md_get_object_dir(oid));
 	} else {
-		snprintf(tree_path, PATH_MAX, "%s/%02x",
+		ret = asprintf(&tree_path, "%s/%02x",
 			 md_get_object_dir(oid), get_tree_directory(oid));
 	}
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	sd_debug("try link %016"PRIx64" from snapshot with epoch %d", oid,
 		 tgt_epoch);
 
-	snprintf(path, sizeof(path), "%s/%016"PRIx64, tree_path, oid);
-	get_store_stale_path(oid, tgt_epoch, 0, stale_path);
+	ret = asprintf(&path, "%s/%016"PRIx64, tree_path, oid);
+	free(tree_path);
+	if (ret < 0)
+		return ret;
+
+	ret = get_store_stale_path(oid, tgt_epoch, 0, &stale_path);
 
 	if (link(stale_path, path) < 0) {
 		/*
@@ -483,9 +566,14 @@ int tree_link(uint64_t oid, uint32_t tgt_epoch)
 			goto out;
 
 		sd_debug("failed to link from %s to %s, %m", stale_path, path);
-		return err_to_sderr(path, oid, errno);
+		free(stale_path);
+		ret = err_to_sderr(path, oid, errno);
+		free(path);
+		return ret;
 	}
 out:
+	free(stale_path);
+	free(path);
 	return SD_RES_SUCCESS;
 }
 
@@ -526,38 +614,57 @@ static int move_object_to_stale_dir(uint64_t oid, const char *wd,
 				    uint32_t epoch, uint8_t ec_index,
 				    struct vnode_info *vinfo, void *arg)
 {
-	char path[PATH_MAX], stale_path[PATH_MAX], tree_path[PATH_MAX - 33];
+	char *path, *stale_path, *tree_path;
 	uint32_t tgt_epoch = *(uint32_t *)arg;
+	int ret;
 
 	if (is_vdi_obj(oid) || is_vmstate_obj(oid) || is_vdi_attr_obj(oid)) {
-		snprintf(tree_path, sizeof(tree_path), "%s/meta",
+		ret = asprintf(&tree_path, "%s/meta",
 			 md_get_object_dir(oid));
 	} else {
-		snprintf(tree_path, sizeof(tree_path), "%s/%02x",
+		ret = asprintf(&tree_path, "%s/%02x",
 			 md_get_object_dir(oid), get_tree_directory(oid));
 	}
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	/* ec_index from md.c is reliable so we can directly use it */
 	if (ec_index < SD_MAX_COPIES) {
-		snprintf(path, sizeof(path), "%s/%016"PRIx64"_%d",
+		ret = asprintf(&path, "%s/%016"PRIx64"_%d",
 			 tree_path, oid, ec_index);
-		snprintf(stale_path, sizeof(stale_path),
+		if (ret < 0) {
+			free(tree_path);
+			return SD_RES_NO_MEM;
+		}
+		ret = asprintf(&stale_path,
 			 "%s/.stale/%016"PRIx64"_%d.%"PRIu32,
 			 md_get_object_dir(oid), oid, ec_index, tgt_epoch);
 	} else {
-		snprintf(path, sizeof(path), "%s/%016" PRIx64,
+		ret = asprintf(&path, "%s/%016" PRIx64,
 			 tree_path, oid);
-		snprintf(stale_path, sizeof(stale_path),
+		if (ret < 0) {
+			free(tree_path);
+			return SD_RES_NO_MEM;
+		}
+		ret = asprintf(&stale_path,
 			 "%s/.stale/%016"PRIx64".%"PRIu32,
 			 md_get_object_dir(oid), oid, tgt_epoch);
 	}
-
+	free(tree_path);
+	if (ret < 0) {
+		free(path);
+		return SD_RES_NO_MEM;
+	}
 	if (unlikely(rename(path, stale_path)) < 0) {
 		sd_err("failed to move stale object %" PRIX64 " to %s, %m", oid,
 		       path);
+		free(stale_path);
+		free(path);
 		return SD_RES_EIO;
 	}
 	sd_debug("moved object %016"PRIx64, oid);
+	free(stale_path);
+	free(path);
 	return SD_RES_SUCCESS;
 }
 
@@ -586,21 +693,28 @@ int tree_format(void)
 
 int tree_remove_object(uint64_t oid, uint8_t ec_index)
 {
-	char path[PATH_MAX];
+	char *path;
+	int ret;
 
 	if (uatomic_is_true(&sys->use_journal))
 		journal_remove_object(oid);
 
-	get_store_path(oid, ec_index, path);
+	ret = get_store_path(oid, ec_index, &path);
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	if (unlink(path) < 0) {
-		if (errno == ENOENT)
+		if (errno == ENOENT) {
+			free(path);
 			return SD_RES_NO_OBJ;
+		}
 
 		sd_err("failed, %s, %m", path);
+		free(path);
 		return SD_RES_EIO;
 	}
 
+	free(path);
 	return SD_RES_SUCCESS;
 }
 
@@ -631,25 +745,39 @@ static int set_object_sha1(const char *path, const uint8_t *sha1)
 	return ret;
 }
 
-static int get_object_path(uint64_t oid, uint32_t epoch, char *path,
-			   size_t size)
+static int get_object_path(uint64_t oid, uint32_t epoch, char **path)
 {
-	char tree_path[PATH_MAX - 33];
+	char *tree_path;
+	int ret;
 
 	if (is_vdi_obj(oid) || is_vmstate_obj(oid) || is_vdi_attr_obj(oid)) {
-	  snprintf(tree_path, sizeof(tree_path), "%s/meta",
+		ret = asprintf(&tree_path, "%s/meta",
 			 md_get_object_dir(oid));
 	} else {
-		snprintf(tree_path, sizeof(tree_path), "%s/%02x",
+		ret = asprintf(&tree_path, "%s/%02x",
 			 md_get_object_dir(oid), get_tree_directory(oid));
 	}
+	if (ret < 0)
+		return SD_RES_NO_MEM;
 
 	if (tree_exist(oid, 0)) {
-		snprintf(path, PATH_MAX, "%s/%016"PRIx64,
+		ret = asprintf(path, "%s/%016"PRIx64,
 			 tree_path, oid);
+		if (ret < 0) {
+			free(tree_path);
+			return SD_RES_NO_MEM;
+		}
 	} else {
-		get_store_stale_path(oid, epoch, 0, path);
-		if (access(path, F_OK) < 0) {
+		ret = get_store_stale_path(oid, epoch, 0, path);
+		if (ret < 0) {
+			free(tree_path);
+			return SD_RES_NO_MEM;
+		}
+			
+		if (access(*path, F_OK) < 0) {
+			free(tree_path);
+			free(*path);
+			*path = NULL;
 			if (errno == ENOENT)
 				return SD_RES_NO_OBJ;
 			return SD_RES_EIO;
@@ -657,6 +785,7 @@ static int get_object_path(uint64_t oid, uint32_t epoch, char *path,
 
 	}
 
+	free(tree_path);
 	return SD_RES_SUCCESS;
 }
 
@@ -667,9 +796,9 @@ int tree_get_hash(uint64_t oid, uint32_t epoch, uint8_t *sha1)
 	struct siocb iocb = {};
 	uint32_t length;
 	bool is_readonly_obj = oid_is_readonly(oid);
-	char path[PATH_MAX];
+	char *path;
 
-	ret = get_object_path(oid, epoch, path, sizeof(path));
+	ret = get_object_path(oid, epoch, &path);
 	if (ret != SD_RES_SUCCESS)
 		return ret;
 
@@ -677,14 +806,17 @@ int tree_get_hash(uint64_t oid, uint32_t epoch, uint8_t *sha1)
 		if (get_object_sha1(path, sha1) == 0) {
 			sd_debug("use cached sha1 digest %s",
 				 sha1_to_hex(sha1));
+			free(path);
 			return SD_RES_SUCCESS;
 		}
 	}
 
 	length = get_store_objsize(oid);
 	ret = posix_memalign((void **)&buf, getpagesize(), length);
-	if (ret)
+	if (ret) {
+		free(path);
 		return SD_RES_NO_MEM;
+	}
 
 	iocb.epoch = epoch;
 	iocb.buf = buf;
@@ -692,6 +824,7 @@ int tree_get_hash(uint64_t oid, uint32_t epoch, uint8_t *sha1)
 
 	ret = tree_read_from_path(oid, path, &iocb);
 	if (ret != SD_RES_SUCCESS) {
+		free(path);
 		free(buf);
 		return ret;
 	}
@@ -704,7 +837,7 @@ int tree_get_hash(uint64_t oid, uint32_t epoch, uint8_t *sha1)
 
 	if (is_readonly_obj)
 		set_object_sha1(path, sha1);
-
+	free(path);
 	return ret;
 }
